Rename ColorMenu to ToolbarMenu and put in own module
